updated the tests on web to perform continuous pings if configured

+++ android/src/main/kotlin/com/shaz/plugin/fist/flutter_internet_speed_test/TestListener.kt
@@ -10,7 +10,7 @@ interface TestListener {
 } 
 
 interface LatencyTestListener {
-    fun onLatencyMeasured(percent: Double, latency: Long, jitter: Double)
+    fun onLatencyMeasured(percent: Double, latency: Double, jitter: Double)
     fun onComplete(averageLatency: Double, jitter: Double)
     fun onError(errorMessage: String)
     fun onCancel()
